[Refactor] `ExportMap`: separate ExportMap instance from its builder logic
